Refactor metric classes
